Alternative implementation to support "use" of drawable components within "defs"
