+++ lib/src/vector_drawable.dart
@@ -4,7 +4,6 @@ import 'dart:ui';
 import 'package:flutter/widgets.dart' hide TextStyle;
 import 'package:meta/meta.dart';
 import 'package:path_drawing/path_drawing.dart';
-import 'package:xml/xml.dart';
 
 import 'render_picture.dart' as render_picture;
 
@@ -281,7 +280,6 @@ class DrawableText implements Drawable {
 class DrawableDefinitionServer {
   final Map<String, PaintServer> _paintServers = <String, PaintServer>{};
   final Map<String, List<Path>> _clipPaths = <String, List<Path>>{};
-  final Map<String, XmlElement> _xmlElements = <String, XmlElement>{};
 
   /// Attempt to lookup a pre-defined [Paint] by [id].
   ///
@@ -307,16 +305,6 @@ class DrawableDefinitionServer {
   void addClipPath(String id, List<Path> paths) {
     assert(id != null);
     _clipPaths[id] = paths;
-  }
-
-  XmlElement getXmlElement(String id) {
-    assert(id != null);
-    return _xmlElements[id];
-  }
-
-  void addXmlElement(String id, XmlElement el) {
-    assert(id != null);
-    _xmlElements[id] = el;
   }
 }
