+++ lib/svg.dart
@@ -99,6 +99,7 @@ class Svg {
     final XmlElement svg = xml.parse(rawSvg).rootElement;
     final Rect viewBox = parseViewBox(svg);
     //final Map<String, PaintServer> paintServers = <String, PaintServer>{};
+    final XmlDefinitionServer xmlDefinitions = new XmlDefinitionServer();
     final DrawableDefinitionServer definitions = new DrawableDefinitionServer();
     final DrawableStyle style = parseStyle(svg, definitions, viewBox, null);
 
@@ -107,6 +108,7 @@ class Svg {
         .map(
           (XmlNode child) => parseSvgElement(
                 child,
+                xmlDefinitions,
                 definitions,
                 viewBox,
                 style,
